Update patch set 1
Patch Set 1:

I think the fault is with the code trying to modify the parameter, and not the code trying to pass Bundle.EMPTY as a parameter. Passing an immutable as a parameter shouldn't be an issue.

But take my comment as a grain of salt, I'm really not familiar with the code around here.

Patch-set: 1