Update patch set 1
Patch Set 1:

The original patch that this reverts causes the incall screen to show the "add call" button even where there is no space for an additional top level call.  We will show the button to add a call, but the call will fail the makeRoomForOutgoingCall check in telecomm so it will be dropped immediately.

QC made the patch (presumably) to allow more calls to be added for DSDA phones... but this is not the right way to do it.  Should revert for now.  

ps this is already reverted/non-existant in 12.1

Patch-set: 1