Update patch set 1
Patch Set 1:

Ok yeah, actually this is the cause right here. Telecom is waiting for the call to be disconnected. This needs to come from QCRIL on certain platforms. However QCRIL is hitting the audio system to get parameters, which it can't do until audio focus is returned, so the whole radio subsystem deadlocks.

Patch-set: 1