Update patch set 1
Patch Set 1: Code-Review-1

This is just a means of artificially keeping a potentially leaky service alive. This is horrible.

If you really just wanted to preempt the service connection, you'd simply call bind() in the InCallController's constructor so that a cached InCallService was readily available on first call. 

Messing with the lifecycle here could cause many issues going forward.

Patch-set: 1