Update patch set 1
Patch Set 1:

Yes, that part isn't necessary to fix my specific bug.  It was however the only place that didn't look like it was consistent.  Basically, I went through the instances of isDsda() and looked for any that were

if (isDsda) {
  if (mBluetoothDsdaService) {
  }
}

and there were only 2 instances.  The one that caused me problems and the other one in this commit.  If you want I can remove it but I suspect it is janky as is.

Patch-set: 1