Update patch set 4
Patch Set 4:

Deepak,

The device I'm working on is the Asus Zenfone2 (Z00A).  It is a DSDA device (obviously) and it is x86 based (so potentially completely different from anything qcom).

The first problem that I had was that if I make a call on sim1 and then make or receive a call on sim2, I end up with no audio on either call.  I can restore the audio by holding and unholding each call (which is effectively what the first version of this commit is doing).  Once you fix that by doing the hold/unhold I end up with other problems depending on what I try to do to fix it properly.

The comment you are asking about is from decompiling the asus stock (5.0) version of this jar to see if I could see anything that they were doing that was obviously different.  They seemed to be using a modified (and also older) version of something based on CAF.  While being based on CAF, it was still fairly different from what we have.  This hold/unhold sequence stood out in the diffs as something that was very obviously different so I played with it.

I can fix the TODO in the commit message with a small change to updateForegroundCall and then the last problem that I have is that when I have 2 calls on sim1 and then initiate a new call on sim2 it does not connect audio on sim2 properly.

However, when I try to fix that I end up making changes that make everything much worse.  That leads me to believe that I am trying to fix some fundamental problem that I don't understand.  Something like QCOM devices implicitly do hold/unhold calls for LCH that aren't done on the intel based modems or something like that.

Do you have any suggestions for me?  (fyi, I'm happy to completely switch approaches if you think something completely different is more likely to be better)

Patch-set: 4