Update patch set 4
Patch Set 4:

I took a break from this for a bit (needed my second sim card in another device) and just came back to it.

It looks like my device completely lacks the concept of "local call hold".  To work around this, it appears that the stock dialer only supports a total of 2 active calls (either 2 on 1 sim or one per sim).

This allows the hold/unhold sequence I added in the first patch set to switch between calls (if because you can only have 1 active call on a sim when switching active subscriptions, the holding of the call acts like the lch by placing one sub on hold and keeping the other active).

Which sounds like I am going to have to do something very device specific in here.

Patch-set: 4