Update patch set 5
Patch Set 5:

But there is no change to the logic with the new variable name. U are defaulting dsdaSupportsLch to true and the new logic is the same, I am not sure what changing the name is buying us in this case. If we were changing the default logic I could understand it, but in this case its not true.

Patch-set: 5