Update patch set 5
Patch Set 5:

Adrian / Deepak,

With this change, we now have 2 classes of devices.  Those with LCH and those without.  Those that support qcom's LCH can have a maximum of 4 top level calls active and those that do not support it can have a maximum of 2 top level calls active at the same time.

That is, this change is changing the existing logic that:

dsda => 4 top level calls
!dsda => 2 top level calls

to the new logic that:

dsda && lch => 4 top level calls
dsda && !lch => 2 top level calls
!dsda => 2 top level calls

If we merge in new code that uses the old name that new code was written with the logic "dsda => 4 top level calls" which is going to be wrong for any device that is DSDA but doesn't support QCOM's LCH (which my device doesn't, probably due to patents).  That's why changing the name is good to let the compiler tell us before some user figures out that the new code is broken on a (admittedly small) subset of devices.

I can rename it if we want to make the choice that we'd rather pick to break less important devices than deal with merge conflicts.  But I do want to make sure we are really making that choice and making it consciously.

Patch-set: 5