Update patch set 3
Patch Set 3:

That's fine. The only requirement from the GSMA documentation (http://www.gsma.com/newsroom/all-documents/ts-11-device-field-and-lab-test-guidelines/) is that the existing call be terminated directly when the new one is picked up.

As long as we kill the Dialog, we should be good. Most of the raciness came from its listeners running in parallel with the callmanager state machine, and operating on Call references that could be either dead, or in a different state.

Still, the original design (from upstream) is still flaky. Calling call1.disconnect() and call2.answer() in parallel can mess with another state machine (the RIL's), so whatever we come up with there should try to serialize the operations.

Patch-set: 3