Update patch set 1
Patch Set 1:

I got the device to show this error on Sambar. I traced it down, the flow is     createAccountHandle(context,sipProfileName)->PhoneAccountHandle(..., sipProfileName). So when you call getUserSelectedVoicePhoneAccount() the id is not a LONG but a string which in this case is the sipprofilename which can be an alpha numeric string. @ethan any comments on how to go about this? I will think the fix is ok it avoids the crash.

Patch-set: 1