Update patch set 1
Patch Set 1:

Also interesting is that in this special case during account registry the function being hit are in phonebase and not sipphone as in the good case. something else is not happening before causing this crash ltr.
PhoneBase: Error! registerForCallWaiting() in PhoneBase should not be called, CDMAPhone inactive.

Patch-set: 1