PhoneAccounts sorting
Any time a PhoneAccount is added to the mState,
perform a sort

Sort order is:
a) Sort CAPABILITY_SIM_SUBSCRIPTION PhoneAccounts
b) Sort by EXTRA_SORT_ORDER
c) Sort by PHoneAccount label.

In case of phone account re-registration in Telecomm,
old account was getting removed and updated account
was getting appended. So, the order of PhoneAccount
was getting changed with the fix this case will
be handled.

Bug: 34872161
Change-Id: I268d2250cc9bcd1a5e9eb8be99dcb23e2bcceadc