Get carrier config value by subId
It uses api "getConfig@CarrierConfigManager" to check whether the
emergency call is allowed in call log or not. But this api just
returns the carrier config of default sub. So the carrier config
may be incorrect if the sub is not the default sub on DS device.

Use "getConfigForSubId@CarrierConfigManager" for different sub.

Test: manual - make an emergency call and check the call log
Bug: 34336505

Change-Id: I1d869ea3e567a51dc766173d9bb7c7391de092ed