Add ringing state for bluetooth audio route
Previously, Telecom was setting up SCO audio at the same time as the
RING command was being sent to the BT headset. This turns out to have
been a violation of the HFP spec, and ag/1199401 started enforcing it,
which caused calls routed over bluetooth to mishandle audio routing.

This change adds a ringing state to the Bluetooth route in CARSM so that
we can be in a bluetooth route with audio focus but without having set
up SCO audio.

Change-Id: I8aabb983ddd4b546d52b10f78929773b823bfbd0
Fix: 30800730