Handle potential emergency numbers properly
In multisim, when prompt is enabled, call to potential emergency
numbers like 91145 which are not actual numbers are not going.
Reason being, for potential emergency numbers,
NewOutgoingCallBroadcastIntentReceiver is not started, that means
no one will set setNewOutgoingCallIntentBroadcastIsDone(). But
since these are not actual emergency prompt will be shown to user,
and when user selects sub, call will not proceed as
setNewOutgoingCallIntentBroadcastIsDone is not set.
Hence set setNewOutgoingCallIntentBroadcastIsDone in case of
potential emergency numbers as part of fix.

Test: manual - Verified the dial to a potential emergency number works
normally after user selects sub.
Bug: 36439795

Change-Id: I4878993ae1b9c9521988a7f47dffc6af8a944ac4