Fix broken System.getStringForUser
This makes work ringtones work as intended when the personal side has
'none' set as a ringtone.

That was failing because this code used to quietly return the answer
for the primary profile no matter what the arguments were, because
the default 3rd argument is UserHandle.myUserId().

We should add a warning against this, or alternatively we should fix
the default argument in the next release after thorough testing.

Bug: 30658854
Change-Id: I7ce702021e45978add17d26244c4f90f8ecdcefc