Update patch set 1
Patch Set 1:

is it just me or is there something totally wrong?

i have telephony-ext on the bootclass path, still all parts depending on it fail to resolve.

results in a soft reboot when i try to call someone.
with this patch, dialer crashes because it can not look up IExtTelephony$Stub

blobs fail as well to resolve it.

when trying to convert dex to jar, it fails to detect the dex version (dex2jar and classyshark)

am i missing something or is this really broken?

Patch-set: 1
Reviewer: Alexander Martinz <8328@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0