Update patch set 1
Patch Set 1: Code-Review+1

Not sure how this groupId stuff (new in 7.1) is supposed to work. The Javadoc of PhoneAccount.getGroupId() says this:

    /**
     * A non-empty {@link String} representing the group that A {@link PhoneAccount} is in or an
     * empty {@link String} if the {@link PhoneAccount} is not in a group. If this
     * {@link PhoneAccount} is in a group, this new {@link PhoneAccount} will replace a registered
     * {@link PhoneAccount} that is in the same group. When the {@link PhoneAccount} is replaced,
     * its user defined defaults and enabled status will also pass to this new {@link PhoneAccount}.
     * Only {@link PhoneAccount}s that share the same {@link ConnectionService} can be replaced.
     *
     * @return A non-empty String Id if this {@link PhoneAccount} belongs to a group.
     * @hide
     */

With the last statement in the @return clause, keeping an empty string seems fine to me. The root cause of the issue looks like the empty string being converted into an empty tag in the XML ("<group_id></group_id>"), for which getText() returns null when reading it back.

Patch-set: 1