Cache default dialer in Telecom (part 1)
In order to avoid a flurry of recursive calls into Telecom on every
Telecom API call, cache the default dialer in the Telecom process and
only update it when either the relevant setting changes or when the
current default dialer has been uninstalled/disabled.

Part 1 of change. Adds the default dialer cache class, but does not
integrate it into the rest of Telecom.

Test: added unit tests
Change-Id: I22225ed1f1b516dc98e80fa4f82c5b72ed289ba0
Merged-In: Ie81e8a052b2463bef5df4c60c443c73db880780b