Add last emergency call time to ConnectionRequest and Call
This can be used to determine if a call is a potential emergency callback.
This cl also renames the EmergencyLocationHelper class to EmergencyCallHelper
since its not just used for location anymore.

This change greatly simplifies the emergency callback handling in the DialerConnectionService
since it will not be able to directly listen for emergency call broadcasts in O.

This cl is dependent on ag/30952524

Bug: 30952524
Test: Added unit test to BasicCallTests and manual testing with real device
Change-Id: I43513c56b92f978f0a70be878a9430aec74ea77f