Fix default phone account handle deserialization
If group id is empty in a default phone account handle, the deserializer
will return null, which means that the user's choice of default outgoing
phone account is not preserved when rebooting. This fixes the
deserialization so that the default phone account handle is initialized
with a empty group id as written instead of not being read from disk at
all.

Change-Id: I7de7b37c22ab4eae8696c7896740adeb35c69f20
Fix: 33538428
Test: manual + added unit test