Cache setDefaultDialer value to fix CTS failures
When CTS tests call setDefaultDialer, they expect the changes to be made
synchronously, because they require these permissions later in the test.
The new DefaultDialerCache didn't update the cache synchronously, instead
it would update the cache when the ContentProvider updated, after a short
delay. This caused CTS tests to be flaky or fail.

I also fixed some of the Telecom Unit Test failures that were occuring.

Test: Run CTS tests.
Bug: 34094508
Change-Id: I6324c84a15c4087ff8acccaf8aae93ea586823d3