Fixing bug with setting the emergency call time.
This cl fixes a bug that caused the connection request extras
to be overwritten when the last emergency call time was added to
the extras.  It also updates the unit test to detect this problem.
It also adds the emergency call time to the call intent extras
instead of to the call extras directly (the unit test was failing
without this change).

This bug causes incoming wifi calls to be missed after an outgoing
emergency call.

Test: updated BasicCallTest and manual testing

Bug: 35323451
Change-Id: I15759daac53d6f6dad53e16e065427944fe3944f