Accept timeout from apps requesting call disconnect
When processing a call disconnected via the ACTION_NEW_OUTGOING_CALL
broadcast, use the timeout specified by the app instead of the default
timeout. This allows apps that may take a bit longer to place a new call
some more time if they need it.

This also changes the default timeout to 500ms and sets a max timeout of
10s for an app-specified timeout.

Bug: 34474757
Test: manual, unit
Merged-In: I3121dcaf973a10043ffcaa0c8251aa21e18d027e
Change-Id: I1b535f17aefffe423cfbc2cb1d42352e8f4b57f0