Support Connection handover between ConnectionServices.
Call:
- added support for initiation of handover on receipt of the
EVENT_REQUEST_HANDOVER call event.
CallAudioManager:
- suppress disconnect tone when handover is in process for a call.
CallsManager:
- when the ConnectionService adds a new incoming call with
EXTRA_IS_HANDOVER, this indicates that the call is being added as the
destination for the handover.  Adding logic to find the ongoing call on
the device which will be handed over to this new call, and confirm that
handover is supported by both ConnectionService.
- on call removal, clean up handover call references.
- When calls change state, handle completion or failure of overall
handover.
- ensure it is possible to add handover calls when there are other ongoing
calls which would normally prevent them from being added.
TestApps:
- Added ability to initiate handover from Test Incall UI.
- Added ability to receive handover from test self-mgd calling app.

Test: Manual
Bug: 37102939
Change-Id: Idfa4325bb1aee34abad5cdb3d8edb48f0186692e