Send PhoneAccount register/unregistered intents to default dialer.
A few cleanups:
1. When sending the register/unregister phone account intents, previously
there was no attempt made to ensure the intents were only sent when a
new account was added, or an existing one was removed.  Cleaned this up
so that we don't over-send the intents.
2. Minor fix to account diff string in PhoneAccountRegistrar which would
cause icon to always be shown as a diff; removed since bitmaps can't be
compared.
3. Also sending the intents to the default dialer app as an explicit
intent.
Cherry-pick from aosp master to resolve conflicts.

Test: Manual
Bug: 37106957
Change-Id: I7fdaa3e225df6fe3aaf292f9b0b93042f69ca469