Fix issue where incoming call shows up while handing over call.
Fixing issue where the "Answering will end your other call" UX shows up
while a handover is in process.  This was a race-condition which could
happen when the handover is marked as success and the incoming handover
call is not yet active.

Solved by tracking in a Call whether it has successfully handed over or
not.  Using this to filter the call out from IncomingCallNotifier.

Test: Manual
Bug: 37768995
Change-Id: I8eeacb88c0c38953f1e4709ea5777c005cbeaf39