Extra security checks in placeCall.
When adding self-managed calls, ensure that the package of the caller
matches the package for the outgoing call phone account, unless the
caller is able to place calls.

This means a self-managed CS can ONLY place calls for its own CS.
However, the default dialer, or other call capable apps (e.g. Auto) can
still place calls on behalf of a self-managed CS.

Test: Created standalone test app to verify cases.
Bug: 38377651
Change-Id: I568b5131ff640ed66c094afae665630772956a17