Fix lack of notification of un-parent on disconnecting conference child.
As a consequence of recent changes to improve efficiency of IMS conferences,
there was a regression when unparenting a child of a conference when it
was disconnected.  We would neglect to send the connection event notifying
the InCallService of the change in parent for the call.

To fix this I split notifying of parent change into a new method which
is called in the right place for the "add existing call" case, and also
from within setParentAndChildCall.  This actually reads nicer and we're
no longer sending this from setChildOf.

Test: Manual, ran CTS, ran unit tests
Bug: 62236503
Change-Id: Ie59a00a39c2ebeee44a81c2db6edf2968fa5b643