Handle Telecom binder death.
There are instances where a ConnectionService dies, but we don't get the
onServiceDisconnected callback.  This is most common in catastrophic
crashes of a ConnectionService.  The result of this is that the calls
associated with that CS get orphaned in Telecom and never cleaned up.

Added a ServiceDeathRecipient implementation to ServiceBinder in Telecom
so that death of a CS is always caught.

Also made some logging changes to ensure we can more easily see when
a CS dies unexpectedly and why its calls were disconnected.

Fixed an issue where DtmfLocalTonePlayer was not finishing its subsession.

Test: Manual
Bug: 62912175
Merged-In: Icbb92dde1792520798f39db7209b5fd672b843c2
Change-Id: I5099592ec781af46c9d123ef2c53f48846b9ec19