Send event when handover source disconnects prior to acceptance of handover.
When the source call disconnects prior to the acceptance of a handover,
send a call event so that the destination connection source is aware
of the fact the source disconnected.  This ensures the destination is
able to either complete the handover (there are cases where a source can
disconnect prior to handover acceptance), or to abort the handover.

Test: Manual using test app
Bug: 62904673
Change-Id: I5b89113160e2697269ecbac7ef4f64844fd8a953