Update patch set 1
Patch Set 1:

(1 comment)

Parsing the id of the phone account is not the best, it's not necessarily the string of an int.

Since we are checking the MCC of the network rather than that of the SIM, I guess using the MCC of the default account is fine. It would have been different if the code used the MCC of the SIM, but this is not the case and all the accounts probably report the same MCC since that's country specific.

Patch-set: 1
Label: Verified=0