Re-enabling temporary location permission granting for emergency calls
We disabled this feature to avoid the case where a user is in a call
then adds an emergency call, then disconnects the emergency call, which
could cause the original call to be disconnected too (because the dialer
process gets killed when location permission is revoked). But that
situation does not currently happen (because adding the emergency
call doesn't go through the EmergencyInCallServiceConnection, and thus
no attempt to grant temporary location permission is made). This cl
fixes that situation by also granting location permission when an
emergency call is added while there is an ongoing call, and revokes
it when all calls have ended.

Note: it is not possible to merge the original call and the emergency call.

Note: the dialer will still be killed after all calls end, if one of them
was an emergency call which was granted temporary location permission.

Bug: 37475749
Test: manual

Change-Id: I12b1362749ee265557d007a989917339468e9b0d