This is a partial cherrypick of ag/3005393
It just contains the fix to location permission granting for emergency
calls for the case where an emergency call is added to an existing call.

The other part of ag/3005393, to enable the feature will be done in
ag/3034149 as a config overlay.

Bug: 67104002
Test: manual
Change-Id: I8c4f308a30b32335718e5909748d3fc84eac7259
Merged-In: I12b1362749ee265557d007a989917339468e9b0d