Mark self-managed calls as "read" when logged.
When self-maanged calls are logged to the call log, ensure that the is_read
flag is set to 1.  This is important for missed calls.  Missed calls which
are unread trigger a missed-call notification.  Realistically we expect
the app to handle the missed called notification on its own, so marking
them as read makes sense.

Also updated the example self managed calling test app to support missed
calls.  Yay.

Test: Added unit tests for this scenario, used manual test harness.
Bug: 67637603
Change-Id: If8488e9ad81f0b4df8d4296e10fb89f3cb56d728