Reconnect to Bluetooth if user switches back to it
Tweaks previous behavior of not auto-connecting to a BT device if the
user explicitly switched off the BT route. Now, if the user switches
back explicitly, we'll auto-reconnect in the future. Add a test for this
too.

Also, make some tweaks to how CARSM updates the rest of the system on
the current audio route and change tests accordingly.

Change-Id: I698fd517007c3c494f3da3343426d51c2b9b5087
Fixes: 66887677
Test: unit