Add connection service focus manager
This changed introduce the ConnectionServiceFocusManager to maintain the
focus status of ConnectionService. When a ConnectionService gained the
focus, it can request the call resource. Also, the ConnectionService
should release the call resource when it lost the focus.

design doc: go/android-telecom-3p-enhancements

Bug: 69651192
Test: unit test

Change-Id: Iea7b4bfd896753ea9d6f399ba341e36150e4e621
Merged-In: Iea7b4bfd896753ea9d6f399ba341e36150e4e621