IMS-VT: Refresh Add call option when Video call downgrades to VoLTE
"android.telecom.extra.DISABLE_ADD_CALL" extra guards the visibility
of "Add Call" option. The option will not be shown if the extra is
present and vice-versa. When call is downgraded, the extra is removed
but is not refreshing the "Add Call" option. Fix this by refreshing
"Add Call" option when the extra is removed.

Test: Manual
Bug: 70128952
Change-Id: Ie4c34e3d681c763009073ec6afee8e2c1fbdb04b