Reset audio route when self-managed calls are disconnected.
When self-managed calls are disconnected to make way for a new managed
call, ensure we switch back to the baseline audio route.  Although audio
routing is considered across all connection services, it is counter
intuitive to the user to be in a speakerphone call (for example) in a
self-managed app and then to have a newly place managed call which
replaces it also route over speakerphone.  We solve this by ensuring that
the audio route switches back to baseline whenever we disconnect all
self-managed calls in order to place a new managed call.  This is
especially important when the user places an emergency call that causes
self-managed calls to disconnect.

Test: Manual test with ril.ecclist; added new unit test to cover this case.
Bug: 70640139
Change-Id: I28100410fe674fe75dc7ddd6d92e968e5970aef4