Add ability to generate remote call recording tone.
Add Telecom support to generate a call recording tone in the outgoing
telephony audio stream when a recording app is enabled on the device.

The new CallRecordingTonePlayer will play a call recording warning tone to
the remote user via outgoing telephony audio stream.  This functionality
if gated on device support and carrier config in the telephony stack.

Test: Added new unit tests.  Manually changed carrier config to enable
tones on local deice, installed audio recording app and verified while it
is recording the tone is audible on remote end.
Bug: 64138141
Change-Id: Ifcf5a49704af5a1352527e705e7d2a3bc16d7fdd