Fix issue where work profiles cause missed calls to not show after boot.
When a user with a work profile reboots their device, Telecom tracks the
user which is pending reload of the missed call notification and defers
the process until boot complete.

This is problematic as there is only one user tracked, and during boot
the work profile user ends up overwriting the primary user.

To fix this, now tracking a list of users to load after boot complete so
that the work profile and primary work profile both reload the missed
call notification, not just the work profile user.

Test: Manually reproduced bug and verified Missed Call notification now
shows up.
Bug: 73164872

Change-Id: I39c87eae9c5dcf85f9d14448dc40acb22c2dc843