Add overlay to control vibration pattern used by device.
Added an overlay which is used to determine which vibration pattern a
device uses.  The new vibration pattern ramps up in intensity, which may
not work well on some vibration hardware.  Devices with vibration hardware
that does not work well with the new pattern can overlay the new config
overlay value to indicate that the old-style simple pattern is used
instead.

Test: Manually tested simple vs fancy vibration pattern.
Bug: 68374140
Change-Id: Icfbac1342de562a2891d7af5d11b79d111f2898d