Take advantage of new audio-coupled haptic patterns.
The new vibrator HAL has a few, device-specific haptic effects that are
tied to a particular piece of audio. The intention is to create custom
physical vibrations that are in sync with a given ringtone, producing a
significantly nicer overall experience.

Making the AsyncRingtonePlayer's Handler truly async reduces the
latency between audio and vibration by 50% to ~5ms based on a few
sample systraces.

Bug: 64185677
Test: atest RingerTest
Change-Id: I0152aa5c8bd5a50563672646dbfa6cc81246a083