Force mute off and speaker on during emergency RTT
In an emergency RTT call, force the speakerphone on at all times. Do not
allow the user to change the audio route to anything other than speaker.

In addition, when any emergency call is in progress, do not allow the
microphone to be muted. If another app attempts to mute the microphone,
immediately unmute it.

Change-Id: I4c6bb7196ddf75161975fb173f8a486dab3f4e6b
Fixes: 74436944
Test: manual