Send CALL intent directly to Telecom CallIntentProcessor.
TelecomManager#placeCall results in a CALL intent being sent from Telecom
to itself.  If there are delays in the system broadcast queue, the CALL
intent could be held up in the queue causing delays in the outgoing call
flow.

Changing the code so we instead just send the intent directly to the
call intent processor.

Test: Manually reproduced broadcast queue spamming issue and verified that
placeCall API does not block in this case any more.
Bug: 76005593

Change-Id: I578fdd56b8ff9542022dd68a4e3a5490307b9a5c