Correct FocusManager issue which prevented swapping calls.
Where a conference call is active and a standalone call is dialed, the
focus manager would not allow swapping between the two calls.
The underlying issue was that the focus manager would try to focus on one
of the child calls of the conference.
Updated the focus manger interface to add an "isFocusable" method which
is used to indicate that external calls and child calls can't be focused.
Added unit test to verify this scenario.

Test: Manually tested conference scenario, added unit test to capture case.
Bug: 74375021
Change-Id: Ie6d1b9957b9dde81dbba0e7ab5dddcf4a61cfd2e