Correct FocusManager issue which prevented swapping calls.
Where a conference call is active and a standalone call is dialed, the
focus manager would not allow swapping between the two calls.
The underlying issue was that the focus manager would try to focus on one
of the child calls of the conference.
Updated the focus manger interface to add an "isFocusable" method which
is used to indicate that external calls and child calls can't be focused.
Added unit test to verify this scenario.

Test: Manually tested conference scenario, added unit test to capture case.
Bug: 74375021

Merged-In: I48cb2b7918117d63996953e5497749d56e27f801
Change-Id: I48cb2b7918117d63996953e5497749d56e27f801
(cherry picked from commit eaff7c48a3ea4bbd35a7679c47823e38fb5f52a8)