Check user restrictions during incoming calls.
Check that DISALLOW_SMS isn't set for a user before activating the
respond-with-sms functionality
Check that DISALLOW_OUTGOING_CALLS isn't set for a user before adding an
incoming call. If it is, silently reject and log the call for the main
user. (Yes, the name is misleading, but the settings UI implies that
it's supposed to be all calls).

Bug: 73804266
Test: manual
Change-Id: Idf32a3fcbe9f8a34fe314bc39d19930e5015acb8
Merged-In: Idf32a3fcbe9f8a34fe314bc39d19930e5015acb8