Ensure phone state broadcast does not include 3p app addresses.
Do not include the address reported by an app implementing the self-mgd
ConnectionService API.  Aothough we want to report the fact we're in
a call using the phone state broadcast, it is not desirable to include
the call identifier information reported by a third party app.  The phone
number is included in the phone state broadcast to help wearable
companion apps and call blocking apps, for example, operate.  Since none
of these apps can operate on self-managed calls it makes no sense to
expose the number to these apps.

Test: Manual test using test app.
Bug: 79343159
Change-Id: I29eca34bba5e1ec636fdc245bb1221b2f63bbcb1