Fix regression in emergency dialing auto-disconnect code.
Change to support the call focus management caused a regression in the
disconnect logic when placing an outgoing call.  Some carriers do not
support holding of calls so the new logic would always disconnect a single
active call.

Reverting to the previous logic which relies on other means to tell if
we have reached the max number of held calls.  In the case of a carrier
which doesn't support call holding, we'll still place the emergency call
while leaving the previous active call alone.

Test: Ensure that single active call is not disconnected when dialing
emergency call.
Test: Ensure that where an active and held call is present, the active
call is disconnected to make room for emergency call.
Fixes: 79381233

Change-Id: I248804fe7a8402fe4a6a9263ea1224a04073e03f