Fix issue with concurrent call disconnect code in Telecom.
The method holdActiveCallForNewCall is used to ensure that an active call
is either held or disconnected when there is a new incoming call.
There is a special case logic block which applies for Telephony calls
where there is an active call and a held call, and a new incoming call
comes in which the user answers.  The special case is intended to
drop the held call so that the current active call can be held so that
the ringing call can be answered.

There was a flaw in that code which was causing the self-managed CS
CTS verifier test to fail.  Essentially the special case logic block was
being invoked, and as a result when there was another active call which
could not be held from a different connection service, the call would be
held instead of disconnected.

Test: Ran CTS Verifier test and ensures it passes.
Test: Manually reproduce the telephony scenario which the special case
block is intended to handle.
Bug: 80407839
Merged-In: I348da7de84f83b1749af41d6985526bcefcc606a
Change-Id: I348da7de84f83b1749af41d6985526bcefcc606a