Fix inability to reject call as secondary user from lock screen.
The rejectCall method in InCallAdapter was performing a check to see if
reply via sms is allowed, however it was doing this outside of the
"clearCallingIdentity" block.  This was causing a crash when trying to
determine if the device is locked.

Test: Reject call as primary user from lock screen.
Test: Reject call as secondary user from lock screen.
Bug: 78463911

Merged-In: I85b5793a1e64c2567df035cfb3d7b6e0daa3181f
Change-Id: I85b5793a1e64c2567df035cfb3d7b6e0daa3181f
(cherry picked from commit a286b7b8f59a59290cdcc384de502b279234876f)