Call redirection service processor in Telecom
A single instance of call redirection processor that handles the call
redirection with user-defined CallRedirectionService and carrier
CallRedirectionService for a single call.

A user-defined call redirection will be performed firstly and a carrier
call redirection will be performed after that; there will be a total of
two call redirection cycles.

Bug: 64959558
Test: Compile (will add unit test later)
Change-Id: I9cb08529853bd63a77e8e68b606c79ae3d399cf4