Try harder to connect to bluetooth
When BluetoothRouteManager is told to connect with a null device or with
a device that doesn't match anything in the availablity list, and the
active device as reported by the BT stack is null, don't just give up.
Try connecting to just any other device.

Bug: 112018166
Test: unit, manual
Change-Id: Iaf6c570e715311c6b263c0373a7ac70e157aaf7b
Merged-In: Iaf6c570e715311c6b263c0373a7ac70e157aaf7b