Fix weird race condition in BT manager
When switching from one BT device to another, sometimes we'll get the
signal that the second device connected right after the first device
disconnected. The transition to the AudioOff state from the
disconnection will sometimes happen at the exact right time so as to
disconnect the second device as well.
This change removes the code that disconnects all remaining devices when
transitioning into AudioOff since it's not actually useful for anything
except creating bugs.

Change-Id: Ieb0a87cc4db990aaa7d9de65d643ade4d9835068
Fixes: 112285293
Test: unit tests