Prepare event-based handovers for deprecation.
Add a toast from the handover initiation to warn the caller of the
deprecation to come.  This will be removed near the end of the Q
development cycle.

Fix a bug in the event-based handover where a destination CS doesn't
create a handover call and can cause a Telecom crash.

Update the telecom test app to support handovers using the public APIs.

Bug: 110846969
Test: Manual using test app.
Change-Id: Icb31216ef0f2a7cf5e6c34975c9eab0b3e5782c9