Ensure VT upgrade requests are auto-rejected when VT is not supported.
We have existing code to ensure that if the phone account doesn't support
VT that incoming VT upgrade requests will be auto-rejected.
Updating this code to account for the case where the phone account still
supports video but the call does not.

Bug: 77937629
Test: Manually verified incoming requests to upgrade to video are rejected
when video is not locally available for a call.
Test: Added unit tests for VideoProviderProxy class.

Change-Id: I9f3203f2e2a59c62781e9b4a40a99c2d9a5c46e4