Add support for network-identified emergency calls in Telecom.
Add support for new PROPERTY_NETWORK_IDENTIFIED_EMERGENCY_CALL which
Telephony can set for calls where the network has identified the call as
being an emergency call.

Add a new system TelecomManager method for determining if there is an
ongoing emergency call (based on dialed number or network identification
of the call).

Test: Added new unit tests in CallsManagerTest
Test: Use new telephony test intent to verify operation of new API.
Bug: 77565333
Change-Id: I0b8c29404b5856b51e953352bbfd5cbbc4483fad