Improve behavior of end call from bluetooth headset.
When ending a call from the BT headset, the existing behavior was to
unhold the background call.  Telecom already unholds the background call
which results in multiple unhold requests for the same call.
Further, the disconnected active call would also have been sent a hold
command due to a bug in CallsManager.

Test: Manual with bluetooth headset.
Bug: 117967146
Change-Id: I89d07ac25a6afb40de5c230442acfd15d3ebe140